requirements.txt
dac>=0.4.3
datasets>=3.1.0
descript_audiotools>=0.7.4
encodec>=0.1.1
librosa>=0.10.2.post1
matplotlib>=3.10.9
nnAudio>=0.3.4
numpy>=2.4.4
pesq>=0.0.4
ptflops>=0.7.4
pyloudnorm>=0.1.1
soundfile>=0.12.1
torch>=2.2.0+cu121
torchaudio>=2.2.0+cu121
tqdm>=4.66.4
transformers>=4.46.3
